fixed many bugs in drive code

+++ drive/DriveBase.hpp
@@ -10,21 +10,21 @@ public:
         : x_mult(1), y_mult(1), rotation_mult(1) {}
 
     //constructor with custom multipliers
-    DriveBase(float x, float y, float rot)
+    DriveBase(double x, double y, double rot)
         : x_mult(x), y_mult(y), rotation_mult(rot) {}
 
-    virtual void drive(float y) {}
-    virtual void drive(float y, float rot) {}
-    virtual void drive(float x, float y, float rot) {}
+    virtual void drive(double y) {}
+    virtual void drive(double y, double rot) {}
+    virtual void drive(double x, double y, double rot) {}
 
     virtual void drive(BetterController* controller) {}
 
-    float getx() { return x_mult; }
-    float gety() { return y_mult; }
-    float getr() { return rotation_mult; }
+    double getx() { return x_mult; }
+    double gety() { return y_mult; }
+    double getr() { return rotation_mult; }
 
 private:
-    float x_mult=0; //parallel to front face of bot
-    float y_mult=0; //perpendicular to front face of bot
-    float rotation_mult=0;
+    double x_mult=0; //parallel to front face of bot
+    double y_mult=0; //perpendicular to front face of bot
+    double rotation_mult=0;
 };
